Reorganize unit and integration tests
