+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = --cov sockeye test -v
+addopts = --cov sockeye test/unit test/integration -v
